🦄 fix https://github.com/otale/tale/issues/72 https://github.com/otale/tale/issues/55 https://github.com/otale/tale/issues/63 https://github.com/otale/tale/issues/69

+++ src/main/java/com/tale/controller/admin/AttachController.java
@@ -17,6 +17,8 @@ import com.tale.controller.BaseController;
 import com.tale.dto.LogActions;
 import com.tale.dto.Types;
 import com.tale.exception.TipException;
+import com.tale.ext.Commons;
+import com.tale.init.TaleConst;
 import com.tale.model.Attach;
 import com.tale.model.Users;
 import com.tale.service.AttachService;
@@ -28,7 +30,9 @@ import org.slf4j.LoggerFactory;
 
 import java.io.File;
 import java.io.IOException;
+import java.util.ArrayList;
 import java.util.Collection;
+import java.util.List;
 import java.util.Map;
 
 /**
@@ -55,6 +59,8 @@ public class AttachController extends BaseController {
                         @QueryParam(value = "limit", defaultValue = "12") int limit) {
         Paginator<Attach> attachPaginator = attachService.getAttachs(new Take(Attach.class).page(page, limit, "id desc"));
         request.attribute("attachs", attachPaginator);
+        request.attribute(Types.ATTACH_URL, Commons.site_option(Types.ATTACH_URL, Commons.site_url()));
+        request.attribute("max_file_size", TaleConst.MAX_FILE_SIZE / 1024);
         return "admin/attach";
     }
 
@@ -68,27 +74,33 @@ public class AttachController extends BaseController {
         Integer uid = users.getUid();
         Map<String, FileItem> fileItemMap = request.fileItems();
         Collection<FileItem> fileItems = fileItemMap.values();
+        List<String> errorFiles = new ArrayList<>();
         try {
-            fileItems.forEach(f -> {
+            fileItems.forEach((FileItem f) -> {
                 String fname = f.fileName();
-                String fkey = TaleUtils.getFileKey(fname);
-                String ftype = TaleUtils.isImage(f.file()) ? Types.IMAGE : Types.FILE;
-                String filePath = TaleUtils.upDir + fkey;
 
-                File file = new File(filePath);
-                try {
-                    Tools.copyFileUsingFileChannels(f.file(), file);
-                    f.file().delete();
-                } catch (IOException e) {
-                    e.printStackTrace();
+                if(f.file().length() / 1024 <= TaleConst.MAX_FILE_SIZE){
+                    String fkey = TaleUtils.getFileKey(fname);
+                    String ftype = TaleUtils.isImage(f.file()) ? Types.IMAGE : Types.FILE;
+                    String filePath = TaleUtils.upDir + fkey;
+
+                    File file = new File(filePath);
+                    try {
+                        Tools.copyFileUsingFileChannels(f.file(), file);
+                        f.file().delete();
+                    } catch (IOException e) {
+                        e.printStackTrace();
+                    }
+                    attachService.save(fname, fkey, ftype, uid);
+                    siteService.cleanCache(Types.C_STATISTICS);
+                } else {
+                    errorFiles.add(fname);
                 }
-                attachService.save(fname, fkey, ftype, uid);
-                siteService.cleanCache(Types.C_STATISTICS);
             });
         } catch (Exception e) {
             return RestResponse.fail();
         }
-        return RestResponse.ok();
+        return RestResponse.ok(errorFiles);
     }
 
     @Route(value = "delete")
@@ -96,9 +108,7 @@ public class AttachController extends BaseController {
     public RestResponse delete(@QueryParam Integer id, Request request) {
         try {
             Attach attach = attachService.byId(id);
-            if (null == attach) {
-                return RestResponse.fail("不存在该附件");
-            }
+            if (null == attach) return RestResponse.fail("不存在该附件");
             attachService.delete(id);
             siteService.cleanCache(Types.C_STATISTICS);
             String upDir = CLASSPATH.substring(0, CLASSPATH.length() - 1);
@@ -106,11 +116,8 @@ public class AttachController extends BaseController {
             logService.save(LogActions.DEL_ARTICLE, attach.getFkey(), request.address(), this.getUid());
         } catch (Exception e) {
             String msg = "附件删除失败";
-            if (e instanceof TipException) {
-                msg = e.getMessage();
-            } else {
-                LOGGER.error(msg, e);
-            }
+            if (e instanceof TipException) msg = e.getMessage();
+            else LOGGER.error(msg, e);
             return RestResponse.fail(msg);
         }
         return RestResponse.ok();
